moving signin logic from controller to service layer
